Added Product creation fixed some bugs

+++ src/main/java/com/example/computerStock/repos/MessageRepo.java
@@ -1,12 +1,11 @@
 package com.example.computerStock.repos;
 
 import com.example.computerStock.domain.Message;
-import com.example.computerStock.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTagAndMessageFor(String tag, String messageFor);
     List<Message> findByMessageFor(String messageFor);
 }
